Squashed commit of the following:
commit f44e783b7338732e53c3b628492ee8aa30cecc31
Merge: 7b851fd daca2ea
Author: theOldSheep <sunnymengyangliu@gmail.com>
Date:   Mon Dec 4 11:49:59 2023 -0700

    Merge branch 'master' into GoldReleaseBoilerTweak

commit 7b851fd89f9d0709588c48adcf91727ca7c6d12f
Author: theOldSheep <sunnymengyangliu@gmail.com>
Date:   Mon Dec 4 11:45:34 2023 -0700

    pipe puzzle tweak and bugfix

commit ea3d955e0a9a51b1be6f708dc23abdbe6c05d648
Author: theOldSheep <sunnymengyangliu@gmail.com>
Date:   Mon Dec 4 00:33:30 2023 -0700

    player move speed multiplier

+++ ChillerUnityProject/Assets/Scripts/Generics/AnimationSpriteClass.cs
@@ -23,7 +23,7 @@ public class AnimationSpriteClass : MonoBehaviour
     // The amount of frames that each give frame should represet
     // if this is 2 then each frame will be shown twice consecutively.
     // i.e. it is now AABBCC instead of ABC
-    public int repetitionFactor = 1;
+    public float repetitionFactor = 1;
 
     protected float currentTime = 0;
     protected int currentFrame = 0;
